Refactored task options dropup
